more core hole updates
git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@693 6419b465-d902-0410-997c-be2266baf723

+++ gpaw/read_setup.py
@@ -125,7 +125,7 @@ http://wiki.fysik.dtu.dk/gpaw/Setups for details."""
             assert self.version >= '0.4'
         if name == 'atom':
             self.Z = int(attrs['Z'])
-            self.Nc = int(attrs['core'])
+            self.Nc = float(attrs['core'])
             self.Nv = int(attrs['valence'])
         elif name == 'xc_functional':
             if attrs['type'] == 'LDA':
